requirements.txt
openai
apache-airflow-providers-google
matplotlib
